Fix check user logic
