+++ src/services/auths.service.ts
@@ -2,11 +2,6 @@ import * as jwt from "jsonwebtoken";
 
 import config from "../config";
 
-interface User {
-  id: number;
-  email: string;
-}
-
 export class AuthService {
   login(user: User) {
     const payload = { ...user };
@@ -15,13 +10,5 @@ export class AuthService {
       audience: "example.com",
       issuer: "example.com",
     });
-  }
-
-  checkUser(token: string, userId: number) {
-    const { id, email } = jwt.decode(token) as User;
-    if (userId !== id) {
-      throw new Error("사용자가 아닙니다.");
-    }
-    return { id, email };
   }
 }
