+++ src/types/users.d.ts
@@ -1,0 +1,4 @@
+interface User {
+  id: number;
+  email: string;
+}
